[CDAP-12902] Fix status of programs across UI
  - Extracts out StatusMapper to be a simple js module
  - Fixes status-mapper in angular to use generic module
  - Updates status of programs in react side to be consistent as pipeline
